Clear jni
