+++ go_module/README.md
@@ -45,11 +45,11 @@ gomobile bind \
   -target=android/arm64 \
   -androidapi=26 \
   -javapkg=com.dobby.gomobile \
-  -o ../kmp_module/backend/build/generated/gomobile/backend.aar \
+  -o ../kmp_module/app/build/generated/gomobile/backend.aar \
   go_module/kotlin_exports
 ```
 
-The Gradle `:backend` module runs this `gomobile bind` step automatically before
+The Gradle `:app` module runs this `gomobile bind` step automatically before
 Android compilation. The generated AAR replaces the previous manual
 `libbackend.so` + JNI bridge.
