--- kmp_module/backend/.gitignore
@@ -1,1 +1,0 @@
-/build
